// packages/react-components/src/types/AsyncValue.ts
/*
 * Copyright 2025 Palantir Technologies, Inc. All rights reserved.
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *     http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

export interface AsyncLoadingValue {
  type: "loading";
  value?: never;
  error?: never;
}

export interface AsyncLoadedValue<V> {
  type: "loaded";
  value: V | undefined;
  error?: never;
}

export interface AsyncReloadingValue<V> {
  type: "reloading";
  value: V | undefined;
  error?: never;
}

export interface AsyncFailedValue<V, E = Error> {
  type: "failed";
  error: E;
  value: V | undefined;
}

export type AsyncValue<V, E = Error> =
  | AsyncLoadingValue
  | AsyncLoadedValue<V>
  | AsyncFailedValue<V, E>;